move resolveReferences to controller_instance since only used there
